refactor(schematics): add schematic for SCSS import migration from @simpl to @siemens
