+++ projects/element-ng/schematics/scss-import-to-siemens-migration/style-mappings.ts
