+++ projects/element-ng/schematics/simpl-siemens-migration/index.ts
@@ -4,8 +4,6 @@
  */
 import { chain, Rule, schematic, SchematicContext, Tree } from '@angular-devkit/schematics';
 
-import { scssMigrationRule } from './scss-migration-rule.js';
-
 // You don't have to export the function as default. You can also have more than one rule factory
 // per file.
 export const simplSiemensMigration = (_options: { path: string }): Rule => {
@@ -13,7 +11,7 @@ export const simplSiemensMigration = (_options: { path: string }): Rule => {
     context.logger.info('🚀 Starting Simpl to Siemens migration...');
     const chainedRules = chain([
       schematic('migrate-ts-imports-to-siemens', _options),
-      scssMigrationRule(_options)
+      schematic('migrate-scss-imports-to-siemens', _options)
     ]);
     return chainedRules(tree, context);
   };
