--- projects/element-ng/schematics/simpl-siemens-migration/mappings/index.ts
@@ -1,5 +1,0 @@
-/**
- * Copyright (c) Siemens 2016 - 2025
- * SPDX-License-Identifier: MIT
- */
-export * from './style-mappings.js';
